+++ src/card-data.js
@@ -1,0 +1,12 @@
+export default [
+  {
+    title: `The Assassination Of Jessie James By The Coward Robert Ford`,
+    rating: 9.8,
+    year: `2018`,
+    duration: `1h&nbsp;13m`,
+    genre: `Comedy`,
+    imgSrc: `./images/posters/accused.jpg`,
+    description: `A priest with a haunted past and a novice on the threshold of her final vows are sent by the Vatican to investigate the death of a young nun in Romania and confront a malevolent force in the form of a demonic nun.`,
+    commentAmount: 13
+  }
+];
