+++ src/filter-data.js
@@ -1,0 +1,33 @@
+
+export default [
+  {
+    caption: `All movies`,
+    href: `#all`,
+    amount: 221,
+    isActive: true
+  },
+  {
+    caption: `Watchlist`,
+    href: `#watchlist`,
+    amount: 15,
+    isActive: false
+  },
+  {
+    caption: `History`,
+    href: `#history`,
+    amount: 33,
+    isActive: false
+  },
+  {
+    caption: `Favorites`,
+    href: `#favorites`,
+    amount: 11,
+    isActive: false
+  },
+  {
+    caption: `Stats`,
+    href: `#stats`,
+    amount: 2,
+    isActive: false
+  }
+];
