+++ src/get-card-element.js
@@ -1,0 +1,22 @@
+export default (title, rating, year, duration, genre, imgSrc, description, commentAmount) => {
+  return `  
+    <article class="film-card">
+      <h3 class="film-card__title">${title}</h3>
+      <p class="film-card__rating">${rating}</p>
+      <p class="film-card__info">
+        <span class="film-card__year">${year}</span>
+        <span class="film-card__duration">${duration}</span>
+        <span class="film-card__genre">${genre}</span>
+      </p>
+      <img src="${imgSrc}" alt="" class="film-card__poster">
+      <p class="film-card__description">${description}</p>
+      <button class="film-card__comments">${commentAmount}comments</button>
+
+      <form class="film-card__controls">
+        <button class="film-card__controls-item button film-card__controls-item--add-to-watchlist"><!--Add to watchlist--> WL</button>
+        <button class="film-card__controls-item button film-card__controls-item--mark-as-watched"><!--Mark as watched-->WTCHD</button>
+        <button class="film-card__controls-item button film-card__controls-item--favorite"><!--Mark as favorite-->FAV</button>
+      </form>
+    </article>
+  `;
+};
