+++ src/utils.js
@@ -1,0 +1,5 @@
+const getRandomNumberInInterval = (min, max) => {
+  return Math.floor(Math.random() * (max - min + 1) + min);
+};
+
+export default getRandomNumberInInterval;
